Added event loop fixes to jquery.ev.js -> lengthen loop time to 5 seconds + 10 for error conditions.  Also, added test to ensure null messages is not hit.
Started to add streaming content handler, but not sure what to put in there.

Updated epochedud to run using Starman.  Need to install starman on plug.  Also, change the location of the content repo to be on apache.  Need to configure this on the plug.

Updated epoch-student to use meta pragma page expires and no-cache.  Not sure if android really honors this.  Doesn't seem like it. :9

So for the teacher, fix the button fadeIn/Out time to fast.  Feels more responsive.  Move the location of start/stop so that it doesn't crush into the top folder in the content list.  Fixed teacher to ignore all # tagged names in messages.