Added JSON2.js.  Should minify this.  It is useful for some JSONy stuff.  If we don't end up needing it I will remove it later.
Commented out the funky jquery.ev.js.  I believe the bug is in some threading in Android OS (not even browser) which may make the JS engine misbehave.  Or maybe the code itself is prone to issues when the timing is off, meaning it isn't thread safe for our purposes.

Implement a dumb long poll based on code found here: http://www.webdevelopmentbits.com/avoiding-long-polling
Loop interval is set to 5 seconds which is SLOWWWW.  I can adjust this based on whether Android can handle it.  Added JSON2.js.  Should minify this.  It is useful for some JSONy stuff.  If we don't end up needing it I will remove it later.

Commented out the funky jquery.ev.js.  I believe the bug is in some threading in Android OS (not even browser) which may make the JS engine misbehave.  Or maybe the code itself is prone to issues when the timing is off, meaning it isn't thread safe for our purposes.

Implement a dumb long poll based on code found here: http://www.webdevelopmentbits.com/avoiding-long-polling  .  Needs testing on Android.  Will commit this little batch and then review the code on Android.  If it behaves, then I will refactor the code into epochedu.js and replace the loop in epoch-teacher.html.  This means we would be able to drop all of the .ev and MXHR code.