Fixed the broken session restore.  Oddly, this won't ever really work ... the server is too quick.  As soon as the brower leaves or reloads, the server gets notified of a user "part" and it cleans up the session.  Perhaps we want to introduce a delay in parting?  I don't know.  It is mostly reliable in forcing the user back to login screen.  However, the button iteself doesn't work after a reload.  probably an event handler function... just need to debug it a bit more.
As for keeping the session alive, we have two problems.  Sessions cannot survive a server crash.  That's ok at the moment.  Also, sessions cannot survive a 'part'.

As for handling failures , say we lose the server and have to reboot, we lose all the local copy of the channels and sessions.  While problematic, it doesn't really hurt anything.  What we must do though is have some kind of server ping, which we check on certain intervals, to validate the sessionid.  If the sessionid is invalidated, we should push the user back to the login screen.