Fixed more crap with the login.  I hate session management.
There was a problem we forgot to add the db files to nodemon-ingore , so every time the db changes, the server reloads.  This is terrible.  Fixed it.  Generally speaking, the /rejoin should work, but I haven't figured out what I am doing wrong.  At the moment, I am writing all sessions to the db.  Also, writing all channels to the db.  Need to think about how to better utilize this.