refs #1 Implement a silly screengrab service via mcpservice
It does what is generally required, displays a png if you access:

http://localhost:8080/screengrab

It's insecure as hell but who cares.  In the future we can implement
some session management and access control.

The current scheme for the screen capture is 1 screen per second, which
is horrible.  But if you consider the common use case, we are trying to
monitor student activity, not screen cast.  The screen cast scenario is
tricky, and is realistically only going to be accomplished when we get
real WebRTC going on mobile.  Otherwise, we should settle for real time
chat, file sharing, remote control presenting, and one way sharing from
desktop to mobile.