refs #8 Fix regression , wasn't upgrading our standalone URLs.  Should be.
This is a trickier problem than it looks.  We always create a payload from
the server route /send, if only to set it as {} .  So by the time the
message gets to our router we have {} set.  It's hard to detect if {} is
empty.  And you can't quickly do an object compare to verify emptiness.
Sigh.  Anyway, fix is easy.  Just check for type === mediaurl.